requirements.txt
beautifulsoup4==4.12.2
Django==4.1.9
django_crispy_forms==2.0
django_localflavor==4.0
django_twilio==0.14.3.1
pdfkit==1.0.0
Pillow==9.5.0
Pillow==9.5.0
Requests==2.31.0
robobrowser==0.5.3
twilio==7.17.0
